backend/requirements.txt
Django>=4.0
djangorestframework
drf-yasg
djangorestframework-simplejwt
gunicorn
whitenoise